requirements.txt
django-summernote
django-embed-video
django-multiselectfield
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
constantly==15.1.0
psycopg2==2.8.6
Django==3.0.4
django-filter==2.2.0
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==7.0.0
whitenoise==5.0.1
